+++ spotbugs/src/main/java/edu/umd/cs/findbugs/config/UserPreferences.java
@@ -345,14 +345,9 @@ public class UserPreferences implements Cloneable {
     public void removeProject(String projectName) {
         // It should only be in list once (usually in slot 0) but check entire
         // list...
-        Iterator<String> it = recentProjectsList.iterator();
-        while (it.hasNext()) {
-            // LinkedList, so remove() via iterator is faster than
-            // remove(index).
-            if (projectName.equals(it.next())) {
-                it.remove();
-            }
-        }
+        // LinkedList, so remove() via iterator is faster than
+        // remove(index).
+        recentProjectsList.removeIf(projectName::equals);
     }
 
     /**