+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Currently the versioning policy of this project follows [Semantic Versioning v2.
 * Reset DataAnalysis.DEBUG back when analysis reaches MAX_ITER
 * Remove unused methods in `BCELUtil`
 * Remove unused methods and deperecated methods in `edu.umd.cs.findbugs.util.Util`
+* Change to removeIf from Iterator and Iterator.remove
 
 ## 4.0.0-beta1 - 2019-03-27